backend/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic[email]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-multipart
Pillow
pyjwt